article_writer/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
customtkinter>=5.0.0

